#2344 created check for proposed changes fields
